rename helper file, update e2e test titles
