+++ src/applications/coronavirus-screener/tests/00.coronavirus-screener.e2e.spec.js
@@ -6,7 +6,7 @@ import {
   visitorScreening,
   staffPass,
   staffScreening,
-} from './screener-test-question-scenarios';
+} from './question-scenario-helper';
 
 // test scenarios for visitors and staff
 export default createE2eTest(client => {
