+++ src/applications/coronavirus-screener/tests/question-scenario-helper.js
